+++ reg_tests/ipolates/sorc/get_input_data.f90
@@ -6,7 +6,7 @@
 ! with a bitmap.
 !------------------------------------------------------------------------
 
- integer(kind=4), allocatable, public :: gdtmpl_input(:)
+ integer,         allocatable, public :: gdtmpl_input(:)
  integer, public                      :: gdtlen_input, gdtnum_input
  integer, public                      :: i_input, j_input
